Add terraform files and remove unnecessary files
